requirements.txt
# DeepSeek API 相关
deepseek==0.1.0
openai==1.3.6

# 网页爬取相关
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# 数据处理相关
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2

# 定时任务相关
APScheduler==3.10.4

# 邮件发送相关
smtplib
email

# PDF处理相关
PyPDF2==3.0.1
pdfplumber==0.10.2

# 工具类
python-dotenv==1.0.0
logging